Import modules and create post request
Import modules needed through flask. Origin for CORS to change to deployed frontend site.
As no database is present, data temporarily stored in an array that resets when code is changed. Once database is set up real information will be input.
GET and POST request tested with frontend form to make sure it is working and recieving data.
Code will need to be refactored once other parts of the frontend and database become available